move authority processing code to a util file
This is to avoid having the various URLs etc in more than one place and
some of the functions are used in more than one place
